Commit before dumping jest for mocha/chai/enzyme
