+++ package.json
@@ -22,19 +22,21 @@
     "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
-    "browser-sync": "^2.12.3",
+    "browser-sync": "^2.12.5",
     "coveralls": "^2.11.9",
     "del": "^2.2.0",
+    "enzyme": "^2.2.0",
     "gulp": "^3.9.1",
     "gulp-autoprefixer": "^3.1.0",
     "gulp-clean-css": "^2.0.6",
     "gulp-sass": "^2.3.1",
     "gulp-uglify": "^1.5.3",
     "jest-cli": "^11.0.2",
-    "ntypescript": "^1.201604232306.1",
+    "ntypescript": "^1.201604262206.1",
     "react": "^15.0.1",
     "react-addons-test-utils": "^15.0.1",
     "react-dom": "^15.0.1",
+    "sinon": "^1.17.3",
     "ts-loader": "^0.8.2",
     "typescript": "^1.9.0-dev.20160423",
     "webpack": "^1.13.0",
@@ -50,12 +52,27 @@
     "moduleFileExtensions": [
       "js",
       "ts",
-      "tsx"
+      "tsx",
+      "json"
     ],
     "unmockedModulePathPatterns": [
-      "<rootDir>/node_modules/react",
-      "<rootDir>/node_modules/react-dom",
-      "<rootDir>/node_modules/react-addons-test-utils"
+        "react",
+        "react-dom",
+        "react-addons-test-utils",
+        "fbjs",
+        "enzyme",
+        "cheerio",
+        "htmlparser2",
+        "underscore",
+        "lodash",
+        "domhandler",
+        "object.assign",
+        "define-properties",
+        "function-bind",
+        "object-keys",
+        "sinon",
+        "es-abstract",
+        "object.values"
     ],
     "verbose": true,
     "collectCoverage": true
