+++ tsconfig.json
@@ -48,16 +48,20 @@
         "lib/js/utils/RISParser.ts",
         "typings/ABT.d.ts",
         "typings/main.d.ts",
+        "typings/main/ambient/enzyme/index.d.ts",
         "typings/main/ambient/jest/index.d.ts",
+        "typings/main/ambient/jsdom/index.d.ts",
         "typings/main/ambient/react-addons-test-utils/index.d.ts",
         "typings/main/ambient/react-dom/index.d.ts",
         "typings/main/ambient/react/index.d.ts",
         "typings/main/ambient/require/index.d.ts",
+        "typings/main/ambient/sinon/index.d.ts",
         "lib/js/components/ImportWindow.tsx",
         "lib/js/components/PubmedWindow.tsx",
-        "lib/js/components/ReferenceWindow/Components/ManualEntryContainer.tsx",
-        "lib/js/components/ReferenceWindow/Components/MetaFields.tsx",
-        "lib/js/components/ReferenceWindow/Components/People.tsx",
+        "lib/js/components/ReferenceWindow/__tests__/People-test.tsx",
+        "lib/js/components/ReferenceWindow/ManualEntryContainer.tsx",
+        "lib/js/components/ReferenceWindow/MetaFields.tsx",
+        "lib/js/components/ReferenceWindow/People.tsx",
         "lib/js/components/ReferenceWindow/ReferenceWindow.tsx",
         "lib/js/PeerReviewMetabox.tsx",
         "lib/js/Reflist.tsx"
