+++ typings.json
@@ -3,10 +3,12 @@
   "version": false,
   "dependencies": {},
   "ambientDependencies": {
+    "enzyme": "registry:dt/enzyme#1.2.0+20160324040416",
     "jest": "registry:dt/jest#0.9.0+20160319033628",
     "react": "registry:dt/react#0.14.0+20160319053454",
     "react-addons-test-utils": "registry:dt/react-addons-test-utils#0.14.0+20160316155526",
     "react-dom": "registry:dt/react-dom#0.14.0+20160316155526",
-    "require": "registry:dt/require#2.1.20+20160316155526"
+    "require": "registry:dt/require#2.1.20+20160316155526",
+    "sinon": "registry:dt/sinon#1.16.0+20160317120654"
   }
 }
